Fixes type hint for positions

+++ stonky/settings.py
@@ -13,7 +13,7 @@ from stonky.types import SortType
 
 @dataclass
 class Settings:
-    positions: Dict[str, int] = field(default_factory=dict)
+    positions: Dict[str, float] = field(default_factory=dict)
     watchlist: List[str] = field(default_factory=list)
     config_path: Path = Path.home() / ".stonky.cfg"
     cash: Dict[str, float] = field(default_factory=dict)
